Convert to Python 3. Add a playlist generator
It has been close to a decade since Python 3 was released. It's about
time to move over completely I think.

Also, I was spoiled a bit by the auto playlists in BeyondPod, so I've
created a simple version of it.

Possible future versions would:
1. Eliminate the need to list every file and only list those that
need to be ordered before everything else.
2. Support a variety of playlist formats. For now, m3u suffices for my
needs.