Rewrite to use requests package
We were receiving 403 statuses from various sites (WOFF and Abject Suffering).
Initially, I thought it had to do with User-Agent (correctly), but even requests
didn't set it by default.

Regardless, the API for requests is much easier to use than what I was trying to
do with the urllib packages. We'll keep it. If we don't like it, we'll revert to
before this commit.