Use isoformat date strings rather than ordinals
They're easier for a human to parse and see when the last check was.
Ironically(?) it was Abject Suffering that prompted this change as I was
downloading many more episodes than I would have expected. Turns out their past
episodes got smooshed into the same day: I'm guessing when they changed feed
providers the date was the date that they converted.

Regardless, using ordinals was hard to track the bug and prompted shifting to
datetime. I did have a few feeds redownload, and I'm not real sure why, but I
don't think it is worth investigating yet.